ENG-8089: Implement Lost+Found with RedisTokenManager (#5927)
* Token manager tracks instance_id in token_to_socket

* RedisTokenManager: keep local dicts globally updated via pub/sub

* Implement lost+found for StateUpdate without websocket

When an update is emitted for a token, but the websocket for that token is on
another instance of the app, post it to the lost+found channel where other
instances are listening for updates to send to their clients.

* Implement `enumerate_tokens` for TokenManager

Set the groundwork for being able to broadcast updates to all connected states.

* Consolidate on `_get_token_owner`

* fix test_connection_banner.py: expect SocketRecord JSON

* Implement real redis-backed test cases for lost+found

* add some polling for the emit mocks since L+F doesn't happen immediately