ENG-8258: API for linking and sharing states (#6024)
* ENG-8258: API for linking and sharing states

It works by defining a substate of SharedState and then calling
self._link_to(target_token) from some event handler.  from that point on,
whenever that user's state is loaded, the StateManager will patch in the linked
shared states.  whenever a linked state is modified, we explicitly load all of
the other linked tokens, patch in the modified states, and send a delta to
those clients

You can call ._unlink to remove the link association, which causes the substate
to be subsequently loaded from the client_token's tree as a private state

It is intended to work transparently with computed vars, background events, and
frontend rendering.

* WiP: move _links to root state to avoid extra redis load

* Fix test_mixin_state

* only import ShareStateBaseInternal if there are links present

* Fix delta calculation in certain scenarios

* Add integration test for linked states

* fix test_computed_var assertion

* remove background task deadlock updating linked state

perform the subsequent updates in an asyncio.Task to allow the original caller
to drop the lock for the other shared states.

* do not reuse _modify_linked_states in `_link_to`

the state might have multiple links, and we may have already entered the
context and are holding the lock already, so we don't want to take the lock
again which will hang.

instead, check if the state is already linked to the target token and avoid
doing extra work.

* abstract out _patch_state and use it when linking or unlinking

* _link_to returns the newly linked state

_modify_linked_states context can now release the locks of newly linked states
and send updates for changes in newly linked states.

rehydrating after linking is no longer necessary.

* fix test cases

* mark router and router_data as dirty when patching state

definitely the token will be different. although private-dependent data should
use private states, it's common for llm generated code to define router_data
dependent vars in the linked state itself, so we make that special case work

* AI CR feedback