ENG-8507: rebind MutableProxy when it is linked to an old _self_state reference
Avoids an issue where MutableProxy references are saved in the state, then when
accessing them again, they reference the wrong StateProxy. So even though the
code enters `async with self`, the StateProxy that is marked mutable is NOT the
StateProxy associated with the value.

With this change, anytime a MutableProxy is accessed through a StateProxy, its
state reference is reset to the current StateProxy before returning it.